Small adjustments to gameplay and some visual changes
Collecting all the souls changes the message when time is over.
Souls give 3 seconds instead of five.

In order to make it clearer that the character isn't able to fly, the flapping of the wings is now faster when ascending and slower when gliding down.

"easy ->" is written at the start of the level to tell the players which direction they should start going.

The player direction is now also controllable with A,D keys.

Collecting all the souls or pressing the G key ends the level and sends the player to the "submit score" screen.